+++ foremoney/transactions/helpers.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+from typing import Iterable, Mapping, Any
+
+
+def make_labels(items: Iterable[Mapping[str, Any]]) -> list[dict[str, Any]]:
+    """Return [{"id": id, "name": "name (value)"}, ...] for items with id,name,value."""
+    return [
+        {"id": item["id"], "name": f"{item['name']} ({item['value']})"}
+        for item in items
+    ]
+
+
+def labels_map(labels: Iterable[Mapping[str, Any]]) -> dict[str, int]:
+    """Map label name to original id."""
+    return {lbl["name"]: lbl["id"] for lbl in labels}
